+++ app/src/main/java/com/example/help_hub/Activities/MainActivity.java
@@ -35,8 +35,6 @@ public class MainActivity extends AppCompatActivity {
         Toolbar myToolbar = findViewById(R.id.my_toolbar);
         setSupportActionBar(myToolbar);
         getSupportActionBar().setDisplayShowTitleEnabled(false);
-        getSupportActionBar().setHomeAsUpIndicator(R.drawable.ic_baseline_arrow_back_24);
-        getSupportActionBar().setDisplayHomeAsUpEnabled(false);
     }
 
     @Override
